:bug: Fix link in contributing (#126)
## Description

The projects all have a `contributing.html` (lower case), while the link
in `CONTRIBUTING.md` was linking to `CONTRIBUTING.html`. This resulted
in a 404 error when clicking the link.

Fixes #125 <!--- Replace `(issue)` with the issue number fixed by this
pull request. If this PR does not fix an issue, please remove this line.
-->

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] ~I have added appropriate tests that cover the new/changed
functionality.~
- [x] ~I have updated the documentation to reflect these changes.~
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] ~I have added migration instructions to the upgrade guide (if
needed).~
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.